services/image-processing-service/src/image_processor.py
"""
Image Processor - Handles image processing operations
"""
from PIL import Image
from io import BytesIO

class ImageProcessor:
    """
    Handles core image processing operations required for analysis
    """

    async def process_image(self, image_data: bytes):
        """
        Process input image data, perform any necessary transformations
        :param image_data: Raw image byte data
        :return: Processed PIL Image object
        """
        # Convert byte data to image
        image = Image.open(BytesIO(image_data))
        # Example transformation
        image = image.convert('RGB')
        
        # Add additional processing if needed
        
        return image

